+++ src/Tracing/GuzzleTracingMiddleware.php
@@ -22,6 +22,7 @@ final class GuzzleTracingMiddleware
         return static function (callable $handler) use ($hub): Closure {
             return static function (RequestInterface $request, array $options) use ($hub, $handler) {
                 $hub = $hub ?? SentrySdk::getCurrentHub();
+                $client = $hub->getClient();
                 $span = $hub->getSpan();
 
                 if (null === $span) {
@@ -34,7 +35,18 @@ final class GuzzleTracingMiddleware
 
                 $childSpan = $span->startChild($spanContext);
 
-                $request = $request->withHeader('sentry-trace', $childSpan->toTraceparent());
+                $request = $request
+                    ->withHeader('sentry-trace', $childSpan->toTraceparent());
+
+                // Check if the request destination is allow listed in the trace_propagation_targets option.
+                if (null !== $client) {
+                    $sdkOptions = $client->getOptions();
+
+                    if (\in_array($request->getUri()->getHost(), $sdkOptions->getTracePropagationTargets())) {
+                        $request = $request
+                            ->withHeader('baggage', $childSpan->toBaggage());
+                    }
+                }
 
                 $handlerPromiseCallback = static function ($responseOrException) use ($hub, $request, $childSpan) {
                     // We finish the span (which means setting the span end timestamp) first to ensure the measured time